OS-2342 client: Object Key Checksums
On an object update/fetch the client will calculate checksums for
A Keys and D Keys which will be checked by the server to verify that no
corruption occurred to the key in transit. On key enumeration, the server
will calculate checksums for keys and the cliente will verify.

- The I/O Descriptor A Key checksum field was change to be a pointer to
reduce the descriptor size when checksums are not enabled.
- Add dkey checksum to appropriate structures

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>