DAOS-3954 vos: not create punch target for non-rebuild case
For non-rebuild case, in spite of for punch object or punch key,
if the target does not exist, it is not necessary to create the
target firstly, then punch it; otherwise, it will leave useless
information in related tree and cause additional incarnation log
and DTX operations.

It also fixes some vos intent issues.

Signed-off-by: Fan Yong <fan.yong@intel.com>